Port changes from XN_BNL branch manually
Due to some diverging branch changes, merging branches was a pain
I've gone through Xuyang's functions from RTS_server_2.prg and
RTS_tools_xn.prg on the XN_BNL branch into the standard files.
There are also some additional vision sequences relating to the
socket calibration. This allows for an extra correction of the
position based on the variation in the socket daughter board
position on the chip.
This does not move over the README files and images though.
This may require some of the images for the vision sequences to be
added to this branch.