Starting to add site selection flexibility
This removes some of the hard coded values of parameters
and instead uses a site.csv file which includes (currently) the
site name and chip type
The SelectSite function then sets some global variables relating
to the robot hand U0, the DF camera offset in the XY plane from the
stinger, and the DF camera focus distance.
Users should copy `site_example.csv` to `site.csv` and change the
fields to match their site.
This file is kept in the .gitignore so it won't keep overwriting
the repo version.
It may be a good idea to add some of the values this sets to the
site.csv file
Next steps are to make sure all the needed site specific points
files are included.