Removes some hard coded values and improves movement commands
Improves the site selection to also include stored offset values
for the down-facing camera. Still requires one hardcoded path
to the site.csv file to get it to work with the main function
but this can be added to the gitignore once uploaded.
There is now a function to calculate and set the offset values too.
Some of the move chip commands have had small tweaks but still
need to look into the rotation commands and any interference with
the vision sequences at the up-facing camera or colision points
inside the sink.
Proceed with caution, instructions will be uploaded to the wiki
soon.