Updates chip orientation handling in move commands
Beginning to replace the current MoveChipFromTrayToSocket and
MoveChipFromSocketToTray to use more generalized functions which
can handle and correct for chip orienation mismatches between
expectation and reality
If you wish to use this functionality, use the
RunMoveChipTrayToSocket
RunMoveChipSocketToTray
commands, which in turn make use of several functions which use
fiducial markers/keys on chips and socket mezzanine boards to
orient things correctly.
Currently, offsets in X and Y and any minor U corrections are
calculated but not implmented
Need to add to the files generated as in the older commands and
make use of the RTS error command
Also needs some site specific corrections
Next - Add the chip analysis for the LArASIC/COLDADC chips
and reintegrate this functionality with the old BNL code which
generates and stores files for corrections
NB These functions will require use of calibrations!
To be added to instructions imminently